+++ core/class/AbeilleCmd.class.php
@@ -11,10 +11,10 @@
             $request2 = $request;
             $eqLogic = $cmd->getEqLogic();
 
-            if (strpos($request2, "#addrGroup#") > 0) {
-                $request2 = str_replace("#addrGroup#", $eqLogic->getConfiguration("Groupe"), $request2);
-                // logMessage('debug', 'request - addGroup : '.$request2);
-            }
+            // if (strpos($request2, "#addrGroup#") > 0) {
+            //     $request2 = str_replace("#addrGroup#", $eqLogic->getConfiguration("Groupe"), $request2);
+            //     // logMessage('debug', 'request - addGroup : '.$request2);
+            // }
 
             // Tcharp38: #GroupeEPx# removed. Now supported thru 'variables' section
             // if (strpos($request2, "#GroupeEP") > 0) {
@@ -24,31 +24,31 @@
             //     // $request .+ "TEST";
             // }
 
-            if (strpos($request2, '#onTime#') > 0) {
-                $onTimeHex = sprintf("%04s", dechex($eqLogic->getConfiguration("onTime") * 10));
-                $request2 = str_replace("#onTime#", $onTimeHex, $request2);
-            }
+            // if (strpos($request2, '#onTime#') > 0) {
+            //     $onTimeHex = sprintf("%04s", dechex($eqLogic->getConfiguration("onTime") * 10));
+            //     $request2 = str_replace("#onTime#", $onTimeHex, $request2);
+            // }
             // logMessage('debug', 'request - onTime: '.$request2);
 
-            if (stripos($request2, '#addrIEEE#') !== false)  {
-                $commandIEEE = $eqLogic->getConfiguration("IEEE", '');
-                if ($commandIEEE != '')
-                    $request2 = str_ireplace('#addrIEEE#', $commandIEEE, $request2);
-            }
-            if (stripos($request2, '#IEEE#') !== false)  {
-                $commandIEEE = $eqLogic->getConfiguration("IEEE", '');
-                if ($commandIEEE != '')
-                    $request2 = str_ireplace('#IEEE#', $commandIEEE, $request2);
-            }
+            // if (stripos($request2, '#addrIEEE#') !== false)  {
+            //     $commandIEEE = $eqLogic->getConfiguration("IEEE", '');
+            //     if ($commandIEEE != '')
+            //         $request2 = str_ireplace('#addrIEEE#', $commandIEEE, $request2);
+            // }
+            // if (stripos($request2, '#IEEE#') !== false)  {
+            //     $commandIEEE = $eqLogic->getConfiguration("IEEE", '');
+            //     if ($commandIEEE != '')
+            //         $request2 = str_ireplace('#IEEE#', $commandIEEE, $request2);
+            // }
             // logMessage('debug', 'request - addrIEEE: '.$request);
 
-            if (stripos($request2, '#ZiGateIEEE#') !== false) {
-                // Logical Id ruche de la forme: Abeille1/0000
-                $rucheIEEE = Abeille::byLogicalId($net.'/0000', 'Abeille')->getConfiguration("IEEE", '');
-                // logMessage('debug', 'Adresse IEEE de la ruche '.$rucheIEEE);
-                if ($rucheIEEE != '')
-                    $request2 = str_ireplace('#ZiGateIEEE#', $rucheIEEE, $request2);
-            }
+            // if (stripos($request2, '#ZiGateIEEE#') !== false) {
+            //     // Logical Id ruche de la forme: Abeille1/0000
+            //     $rucheIEEE = Abeille::byLogicalId($net.'/0000', 'Abeille')->getConfiguration("IEEE", '');
+            //     // logMessage('debug', 'Adresse IEEE de la ruche '.$rucheIEEE);
+            //     if ($rucheIEEE != '')
+            //         $request2 = str_ireplace('#ZiGateIEEE#', $rucheIEEE, $request2);
+            // }
             // logMessage('debug', 'request - ZigateIEEE: '.$request2);
 
             // Request with multi inputs, input from a Info command.
@@ -68,41 +68,41 @@
             }
             // log::add( 'Abeille', 'debug', 'CmdInfo: End analysis' );
 
-            if (isset($_options)) {
-                switch ($cmd->getSubType()) {
-                case 'slider':
-                    $sliderVal = $_options['slider'];
-                    $sliderVal = trim($sliderVal); // Remove potential space seen at end of slider value
-                    $cmdTopic = $cmd->getConfiguration('topic', '');
-                    if ($cmdTopic == "writeAttribute") {
-                        // New way of handling #slider#
-                        $request2 = str_replace('#slider#', '#slider'.$sliderVal.'#', $request2);
-                    } else
-                        $request2 = str_replace('#slider#', $sliderVal, $request2);
-                    break;
-                case 'select':
-                    $sliderVal = $_options['select'];
-                    $sliderVal = trim($sliderVal); // Remove potential space seen at end of slider value
-                    $cmdTopic = $cmd->getConfiguration('topic', '');
-                    if ($cmdTopic == "writeAttribute") {
-                        // New way of handling #slider#
-                        $request2 = str_replace('#select#', '#select'.$sliderVal.'#', $request2);
-                    } else
-                        $request2 = str_replace('#select#', $sliderVal, $request2);
-                    break;
-                case 'color':
-                    $request2 = str_replace('#color#', $_options['color'], $request2);
-                    break;
-                case 'message':
-                    if (isset($_options['title']))
-                        $request2 = str_replace('#title#', $_options['title'], $request2);
-                    if (isset($_options['message']))
-                        $request2 = str_replace('#message#', $_options['message'], $request2);
-                    break;
-                default: // 'other'
-                    break;
-                }
-            }
+            // if (isset($_options)) {
+            //     switch ($cmd->getSubType()) {
+            //     case 'slider':
+            //         $sliderVal = $_options['slider'];
+            //         $sliderVal = trim($sliderVal); // Remove potential space seen at end of slider value
+            //         $cmdTopic = $cmd->getConfiguration('topic', '');
+            //         if ($cmdTopic == "writeAttribute") {
+            //             // New way of handling #slider#
+            //             $request2 = str_replace('#slider#', '#slider'.$sliderVal.'#', $request2);
+            //         } else
+            //             $request2 = str_replace('#slider#', $sliderVal, $request2);
+            //         break;
+            //     case 'select':
+            //         $sliderVal = $_options['select'];
+            //         $sliderVal = trim($sliderVal); // Remove potential space seen at end of slider value
+            //         $cmdTopic = $cmd->getConfiguration('topic', '');
+            //         if ($cmdTopic == "writeAttribute") {
+            //             // New way of handling #slider#
+            //             $request2 = str_replace('#select#', '#select'.$sliderVal.'#', $request2);
+            //         } else
+            //             $request2 = str_replace('#select#', $sliderVal, $request2);
+            //         break;
+            //     case 'color':
+            //         $request2 = str_replace('#color#', $_options['color'], $request2);
+            //         break;
+            //     case 'message':
+            //         if (isset($_options['title']))
+            //             $request2 = str_replace('#title#', $_options['title'], $request2);
+            //         if (isset($_options['message']))
+            //             $request2 = str_replace('#message#', $_options['message'], $request2);
+            //         break;
+            //     default: // 'other'
+            //         break;
+            //     }
+            // }
             // logMessage('debug', 'request - options: '.$request2);
 
             $request2 = str_replace('\\', '', jeedom::evaluateExpression($request2));
@@ -135,11 +135,13 @@
 
             /* If cmd sub type is not 'other', there is a value associated to cmd which depends on sub type
                'select' => #select#
+                    Can be overloaded by valueOffset
                'slider' => #slider#
+                    Can be overloaded by valueOffset
                'message' => #title# (optional) + #message#
                'color' => #color#
-
-               Value can be overloaded or updated thanks to 'valueOffset' for cases above except 'other'
+               'other' => NO given value
+                    Can use #valueoffset#
              */
 
             // TODO: A revoir, je ne sais plus ce qu'est ce truc.
@@ -158,7 +160,7 @@
             $eqLogicId = $eqLogic->getLogicalId();
             list($net, $addr) = explode("/", $eqLogicId);
             if ($net == '' || $addr == '') {
-                logMessage('error', $eqLogic->getHumanName().': DB corrompue. Cmde annulée.');
+                logMessage('error', $eqLogic->getHumanName().': Cet équipement à besoin de réparation.');
                 return;
             }
 
@@ -191,108 +193,97 @@
                 $request = $cmdLogic->getConfiguration('request', '');
             }
 
-            // Value update if 'valueOffset' is defined.
-            // Reminder: 'valueOffset' is equivalent to 'calculValueOffset' for action cmd.
-            // Ex: "valueOffset": "#value#*10"
-            // Ex: "valueOffset": "#value#*#logicid0102-01-F003#/100"
-            // Ex: "valueOffset": "#valueformat-%02X#" (equiv 'sprintf("%02X", #value#)')
-            $vo = $cmdLogic->getConfiguration('ab::valueOffset', "");
-            if ($vo != "") {
-                logMessage('debug', "-- 'ab::valueOffset'=${vo}");
-                // Replacing '#logicid...#
-                $lop = strpos($vo, '#logicid'); // Any #logicid....# variable ?
-                if ($lop !== false) {
-                    // logMessage('debug', "-- execute(): logicId at pos ".$lop);
-                    $logicId = substr($vo, $lop + 8);
-                    $lop = strpos($logicId, '#');
-                    $logicId = substr($logicId, 0, $lop);
-                    // logMessage('debug', "-- execute(): logicId=".$logicId);
-                    $cmdLogic2 = $eqLogic->getCmd('info', $logicId);
-                    if (!is_object($cmdLogic2)) {
-                        message::add("Abeille", "${cmdHName}: Commande '${logicId}' inconnue");
-                    } else {
-                        $cmdVal = $cmdLogic2->execCmd();
-                        logMessage('debug', "-- Cmd logicId='".$logicId."', val=".$cmdVal);
-                        $vo = str_replace('#logicid'.$logicId.'#', $cmdVal, $vo); // Replace #logicid...#
-                    }
-                }
-
-                if (isset($_options['slider'])) {
-                    $vo = str_replace('#value#', $_options['slider'], $vo); // Replace #value# by slider value
-                    // $lop = strpos($vo, '#logicid'); // Any #logicid....# variable ?
-                    // if ($lop != false) {
-                    //     // logMessage('debug', "-- execute(): logicId at pos ".$lop);
-                    //     $logicId = substr($vo, $lop + 8);
-                    //     $lop = strpos($logicId, '#');
-                    //     $logicId = substr($logicId, 0, $lop);
-                    //     // logMessage('debug', "-- execute(): logicId=".$logicId);
-                    //     $cmdLogic2 = $eqLogic->getCmd('info', $logicId);
-                    //     if (!is_object($cmdLogic2)) {
-                    //         message::add("Abeille", $eqLogic->getHumanName().": Commande '".$logicId."' inconnue");
-                    //     } else {
-                    //         $cmdVal = $cmdLogic2->execCmd();
-                    //         logMessage('debug', "-- Cmd logicId='".$logicId."', val=".$cmdVal);
-                    //         $vo = str_replace('#logicid'.$logicId.'#', $cmdVal, $vo); // Replace #logicid...#
-                    //     }
-                    // }
-                    $newValue = jeedom::evaluateExpression($vo); // Compute final formula
-                    logMessage('debug', "-- 'valueOffset' applied: ".$_options['slider']." => ".$newValue);
-                    $_options['slider'] = $newValue;
-                } else if (strpos($request, '#value#') !== false) {
-                    // #value# must be replaced by valueOffset result
-                    // $lop = strpos($vo, '#logicid'); // Any #logicid....# variable ?
-                    // if ($lop != false) {
-                    //     // logMessage('debug', "-- execute(): logicId at pos ".$lop);
-                    //     $logicId = substr($vo, $lop + 8);
-                    //     $lop = strpos($logicId, '#');
-                    //     $logicId = substr($logicId, 0, $lop);
-                    //     // logMessage('debug', "-- execute(): logicId=".$logicId);
-                    //     $cmdLogic2 = $eqLogic->getCmd('info', $logicId);
-                    //     if (!is_object($cmdLogic2)) {
-                    //         message::add("Abeille", $eqLogic->getHumanName().": Commande '".$logicId."' inconnue");
-                    //     } else {
-                    //         $cmdVal = $cmdLogic2->execCmd();
-                    //         logMessage('debug', "-- Cmd logicId='".$logicId."' => val=".$cmdVal);
-                    //         $vo = str_replace('#logicid'.$logicId.'#', $cmdVal, $vo); // Replace #logicid...#
-                    //     }
-                    // }
-                    $newValue = jeedom::evaluateExpression($vo); // Compute final formula
-                    $request = str_replace('#value#', $newValue, $request);
-                    logMessage('debug', "-- 'valueOffset' applied: newValue=${newValue}");
-                }
-
-                if (isset($_options['slider']))
-                    $value = $_options['slider'];
-                else if (isset($_options['select']))
-                    $value = $_options['select'];
-
-                // #valueformat-FMT# ?
-                $vFmtPos = stripos($vo, '#valueformat-'); // Any #valueformat-FMT# variable ?
-                if ($vFmtPos !== false) {
-                    $fmt = substr($vo, $vFmtPos + 13); // 'FMT#'
-                    $end = strpos($fmt, '#'); // '#' position
-                    $fmt = substr($fmt, 0, $end); // 'FMT'
-                    $newValue = sprintf($fmt, $value);
-                    $request = str_ireplace("#valueformat-${fmt}#", $newValue, $request);
-                    logMessage('debug', "-- 'valueformat': Replaced '${value}' by '${newValue}'");
-                }
-            }
-
-            // -------------------------------------------------------------------------
-            // Process topic
-            // Needed for Telecommande: "topic":"CmdAbeille\/#addrGroup#\/OnOffGroup"
-            if (strpos($cmdLogic->getConfiguration('topic'), "CmdAbeille") === 0) {
-                $topic = str_replace("Abeille", $net, $cmdLogic->getConfiguration('topic'));
-            } else {
-                $topic = "Cmd".$eqLogicId."/".$cmdLogic->getConfiguration('topic');
-            }
-
             /* 'topic' replacement examples
              "topic": "CmdAbeille/#addrGroup#/setColourGroup" => must replace #ADDRGROUP#
              "topic": "CmdAbeille/#GROUPEP4#/OnOffGroup"  => must replace #GROUPEPx#
              "topic": "CmdAbeille/#addrGroup#/setTemperatureGroup
             */
+            if (strpos($cmdLogic->getConfiguration('topic'), "CmdAbeille") === 0) {
+                $topic = str_replace("Abeille", $net, $cmdLogic->getConfiguration('topic'));
+            } else {
+                $topic = "Cmd".$eqLogicId."/".$cmdLogic->getConfiguration('topic');
+            }
+            if (stripos($topic, "#addrGroup#") !== false) {
+                $topic = str_ireplace("#addrGroup#", $eqLogic->getConfiguration("Groupe"), $topic);
+                logMessage('debug', "-- topic: '#addrgroup#' replaced => topic=${topic}");
+            }
             $topic = $cmdLogic->updateField($net, $cmdLogic, $topic, $_options);
+
+            // Value update if 'valueOffset' is defined.
+            // Reminder: 'valueOffset' is equivalent to 'calculValueOffset' for action cmd.
+            // Ex: "valueOffset": "#value#*10" >>> OBSOLETE
+            // Ex: "valueOffset": "#value#*#logicid0102-01-F003#/100" >>> OBSOLETE
+            // Ex: "valueOffset": "#valueformat-%02X#" (equiv 'sprintf("%02X", #value#)')
+            // Ex: 'valueOffset': '100-#slider#'
+            $vo = $cmdLogic->getConfiguration('ab::valueOffset', "");
+            if ($vo != "") {
+                // logMessage('debug', "-- 'valueOffset' = '${vo}'");
+
+                // Replacing '#logicid...#' in valueOffset
+                $pos = stripos($vo, '#logicid'); // Any #logicid....# variable ?
+                if ($pos !== false) {
+                    // logMessage('debug', "-- execute(): logicId at pos ".$pos);
+                    $logicId = substr($vo, $pos + 8);
+                    $lop = strpos($logicId, '#');
+                    $logicId = substr($logicId, 0, $pos);
+                    // logMessage('debug', "-- execute(): logicId=".$logicId);
+                    $cmdLogic2 = $eqLogic->getCmd('info', $logicId);
+                    if (!is_object($cmdLogic2)) {
+                        logMessage("error", "${cmdHName}: Commande '${logicId}' inconnue");
+                    } else {
+                        $cmdVal = $cmdLogic2->execCmd();
+                        logMessage('debug', "-- Cmd logicId='".$logicId."', val=".$cmdVal);
+                        $vo0 = $vo;
+                        $vo = str_ireplace('#logicid'.$logicId.'#', $cmdVal, $vo); // Replace #logicid...#
+                        logMessage('debug', "-- valueOffset: '#logicid${logicId}#' replaced: '${vo0}' => '${vo}'");
+                    }
+                }
+
+                // Replacing '#slider#' in valueOffset
+                $pos = stripos($vo, '#slider#'); // Any #slider# variable ?
+                if ($pos !== false) {
+                    if (!isset($_options['slider'])) {
+                        logMessage("error", "${cmdHName}: 'valueOffset' avec '#slider#' mais cmd pas de type 'slider'");
+                        return;
+                    }
+                    $vo0 = $vo;
+                    $vo = str_ireplace('#slider#', $_options['slider'], $vo); // Replace #slider# by slider value
+                    logMessage('debug', "-- valueOffset: '#slider#' replaced: '${vo0}' => '${vo}'");
+                }
+
+                // Replacing '#valueformat-FMT#' in valueOffset
+                $pos = stripos($vo, '#valueformat-'); // Any #valueformat-FMT# variable ?
+                if ($pos !== false) {
+                    if (isset($_options['slider']))
+                        $value = $_options['slider'];
+                    else if (isset($_options['select']))
+                        $value = $_options['select'];
+                    else {
+                        logMessage("error", "${cmdHName}: 'valueOffset' avec '#valueformat..#' mais mauvais type.");
+                        return;
+                    }
+
+                    $fmt = substr($vo, $pos + 13); // 'FMT#'
+                    $end = strpos($fmt, '#'); // '#' position
+                    $fmt = substr($fmt, 0, $end); // 'FMT'
+                    $newValue = sprintf($fmt, $value);
+                    $vo0 = $vo;
+                    $vo = str_ireplace("#valueformat-${fmt}#", $newValue, $vo);
+                    logMessage('debug', "-- valueOffset: '#valueformat-${fmt}#' replaced: '${vo0}' => '${vo}'");
+                }
+
+                if (isset($_options['slider'])) {
+                    $vo = str_ireplace('#value#', $_options['slider'], $vo); // Replace #value# by slider value
+                }
+
+                // Updating 'request' with valueOffset result
+                $voValue = jeedom::evaluateExpression($vo); // Compute final formula
+                // if (stripos($request, '#valueoffset#') !== false)
+                //     $request = str_ireplace('#valueoffset#', $newValue, $request);
+                // else if (stripos($request, '#value#') !== false)
+                //     $request = str_ireplace('#value#', $newValue, $request);
+                logMessage('debug', "-- valueOffset: result=${voValue}");
+            }
 
             /* 'request' replacement examples
              "request": "Level=#slider#&duration=01"
@@ -303,22 +294,92 @@
              "request": "targetExtendedAddress=#addrIEEE#&targetEndpoint=07&clusterID=0008&reportToGroup=#GROUPEP7#"
              "request": "ep=#EP#&clustId=#CLUSTID#&attrId=#ATTRID#&attrVal=#select#"
              "request": "addr=#IEEE#&ep=#EP#&clustId=#CLUSTID#&destAddr=#ZigateIEEE#&destEp=01"
+             "request": "action=Off&onTime=#onTime#&offWaitTime=0000"
             */
-            if (stripos($request, "#valueoffset#") !== false) {
-                if ($vo == "") {
-                    // message::add("Abeille", "${cmdHName}: 'valueOffset' manquant");
-                    logMessage('error', "${cmdHName}: 'valueOffset' manquant");
+            // if (stripos($request, "#valueoffset#") !== false) {
+            //     if ($vo == "") {
+            //         // message::add("Abeille", "${cmdHName}: 'valueOffset' manquant");
+            //         logMessage('error', "${cmdHName}: 'valueOffset' manquant");
+            //         return;
+            //     }
+            //     /* Note: could not make it work with jeedom::evaluateExpression()
+            //     [2024-06-16 22:05:20] -- Cmd logicId='0102-01-0017', val=0
+            //     [2024-06-16 22:05:20] -- '#valueoffset#' => newValue='0|(1<<1)'
+            //     [2024-06-16 22:05:20] -- '#valueoffset#' => request='ep=01&clustId=0102&attrId=0017&attrVal=0|(1<<1)&attrType=18'
+            //     */
+            //     $exp = "\$newVal = ${vo};";
+            //     eval($exp);
+            //     $request = str_ireplace("#valueoffset#", $newVal, $request);
+            //     logMessage('debug', "-- '#valueoffset#' => newVal='${newVal}', request='${request}'");
+            // }
+            if (stripos($request, "#onTime#") !== false) {
+                $onTimeHex = sprintf("%04s", dechex($eqLogic->getConfiguration("onTime") * 10));
+                $request = str_ireplace("#onTime#", $onTimeHex, $request);
+                logMessage('debug', "-- request: '#onTime#' replaced => request='${request}'");
+            }
+            if (stripos($request, "#addrGroup#") !== false) {
+                $request = str_ireplace("#addrGroup#", $eqLogic->getConfiguration("Groupe"), $request);
+                logMessage('debug', "-- request: '#addrgroup#' replaced => request='${request}'");
+            }
+            if (stripos($request, "#ZiGateIEEE#") !== false) {
+                $ieee = Abeille::byLogicalId($net.'/0000', 'Abeille')->getConfiguration("IEEE", '');
+                $request = str_ireplace("#ZiGateIEEE#", $ieee, $request);
+                logMessage('debug', "-- request: '#ZiGateIEEE#' replaced => request='${request}'");
+            }
+            if (stripos($request, '#IEEE#') !== false)  {
+                $ieee = $eqLogic->getConfiguration("IEEE", '');
+                $request = str_ireplace('#IEEE#', $ieee, $request);
+                logMessage('debug', "-- request: '#IEEE#' replaced => request='${request}'");
+            }
+            if (stripos($request, '#addrIEEE#') !== false)  {
+                $ieee = $eqLogic->getConfiguration("IEEE", '');
+                $request = str_ireplace('#addrIEEE#', $ieee, $request);
+                logMessage('debug', "-- request: '#addrIEEE#' replaced => request='${request}'");
+            }
+            switch ($cmdLogic->getSubType()) {
+            case 'slider':
+                if (!isset($_options['slider'])) {
+                    logMessage("error", "${cmdHName}: 'slider' mais pas de valeur associée.");
                     return;
                 }
-                /* Note: could not make it work with jeedom::evaluateExpression()
-                [2024-06-16 22:05:20] -- Cmd logicId='0102-01-0017', val=0
-                [2024-06-16 22:05:20] -- '#valueoffset#' => newValue='0|(1<<1)'
-                [2024-06-16 22:05:20] -- '#valueoffset#' => request='ep=01&clustId=0102&attrId=0017&attrVal=0|(1<<1)&attrType=18'
-                */
-                $exp = "\$newVal = ${vo};";
-                eval($exp);
-                $request = str_ireplace("#valueoffset#", $newVal, $request);
-                logMessage('debug', "-- '#valueoffset#' => newVal='${newVal}', request='${request}'");
+                $sliderVal = ($vo != "") ? $voValue : $_options['slider'];
+                $sliderVal = trim($sliderVal); // Remove potential space seen at end of slider value
+                $cmdTopic = $cmdLogic->getConfiguration('topic', '');
+                if ($cmdTopic == "writeAttribute") {
+                    // New way of handling #slider#
+                    $request = str_ireplace('#slider#', '#slider'.$sliderVal.'#', $request);
+                } else
+                    $request = str_ireplace('#slider#', $sliderVal, $request);
+                break;
+            case 'select':
+                if (!isset($_options['select'])) {
+                    logMessage("error", "${cmdHName}: 'select' mais pas de valeur associée.");
+                    return;
+                }
+                $sliderVal = ($vo != "") ? $voValue : $_options['select'];
+                $sliderVal = trim($sliderVal); // Remove potential space seen at end of slider value
+                $cmdTopic = $cmdLogic->getConfiguration('topic', '');
+                if ($cmdTopic == "writeAttribute") {
+                    // New way of handling #slider#
+                    $request = str_ireplace('#select#', '#select'.$sliderVal.'#', $request);
+                } else
+                    $request = str_ireplace('#select#', $sliderVal, $request);
+                break;
+            case 'color':
+                $request = str_replace('#color#', $_options['color'], $request);
+                break;
+            case 'message':
+                if (isset($_options['title']))
+                    $request = str_replace('#title#', $_options['title'], $request);
+                if (isset($_options['message']))
+                    $request = str_replace('#message#', $_options['message'], $request);
+                break;
+            default: // 'other'
+                if (stripos($request, "#valueoffset#") !== false) {
+                    $request = str_ireplace("#valueoffset#", $voValue, $request);
+                    logMessage('debug', "-- '#valueoffset#' replaced to request => request='${request}'");
+                }
+                break;
             }
             $request = $cmdLogic->updateField($net, $cmdLogic, $request, $_options);